refactor: separate data logic
