+++ src/app/components/Modal/ModalDetails.js
